process_fsm: fix race condition in start_process
@process was assigned inside a different fiber than the event loop.
THE RACE CONDITION
process.start
 - state: STOPPED
 - Event::START -> State::STARTING -> spawn { start_process }(no fiber
   switch here)
process.stop
 - Event::STOP -> process.kill ( but since @process is nil...so no
    signal sent ) -> State::STOPPING
 - start_process gets executed and @process gets assigned

this results in an inconsistent state where process is running(and not
signalled to stop) but the state is State::STOPPING. Also the "process.stop"
call hangs until the child process is manually killed.

THE FIX
(@process = Process.new) is now executed inside fsm event loop fiber.
(@process = nil) is still run in a different fiber as it doesn't cause any
behavioural issues.